% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool-residuals.R
\name{lavResiduals.mi}
\alias{lavResiduals.mi}
\title{Covariance and Correlation Residuals}
\usage{
lavResiduals.mi(object, omit.imps = c("no.conv", "no.se"), ...)
}
\arguments{
\item{object}{An object of class \code{lavaan.mi}}

\item{omit.imps}{\code{character} indicating criteria for excluding imputations
from pooled results. See \linkS4class{lavaan.mi} for argument details.}

\item{\dots}{Arguments passed to \code{\link[lavaan:lavResiduals]{lavaan::lavResiduals()}}.}
}
\value{
A \code{list} of residuals and other information (see \code{\link[lavaan:lavResiduals]{lavaan::lavResiduals()}}).
}
\description{
This function calculates residuals for sample moments (e.g., means and
(co)variances, means) from a lavaan model fitted to multiple imputed data
sets, along with summary and inferential statistics about the residuals.
}
\examples{

data(HS20imps) # import a list of 20 imputed data sets

## specify CFA model from lavaan's ?cfa help page
HS.model <- '
  visual  =~ x1 + x2 + x3
  textual =~ x4 + x5 + x6
  speed   =~ x7 + x8 + x9
'
## fit model to 20 imputed data sets
fit <- cfa.mi(HS.model, data = HS20imps)

## default type = "cor.bentler" (standardized covariance residuals)
lavResiduals.mi(fit, zstat = FALSE)
## SRMR is in the $summary

## correlation residuals
lavResiduals.mi(fit, zstat = FALSE, type = "cor")
## CRMR is in the $summary

## raw covariance residuals
lavResiduals.mi(fit, type = "raw") # zstat=TRUE by default
## RMR is in the $summary
## "normalized" residuals are in $cov.z

}
\seealso{
\code{\link[lavaan:lavResiduals]{lavaan::lavResiduals()}} for details about other arguments.
}
\author{
Terrence D. Jorgensen (University of Amsterdam;
\email{TJorgensen314@gmail.com})
}
